+++ CHANGELOG.rst
@@ -6,8 +6,10 @@ This document describes changes between each past release.
 1.2.0 (unreleased)
 ------------------
 
-- Nothing changed yet.
+**New features**
 
+- Kinto Admin UI fields like ``displayFields`` ``attachment`` and ``sort`` are copied
+  from the source to the preview and destination collections (if not set) (fixes #161)
 
 1.1.1 (2017-01-17)
 ------------------